+++ ALTAIR/altair_sys.c
@@ -52,7 +52,7 @@ char sim_name[] = "Altair 8800";
 
 REG *sim_PC = &cpu_reg[0];
 
-int32 sim_emax = 4;
+size_t sim_emax = 4;
 
 DEVICE *sim_devices[] = {
     &cpu_dev,
