+++ HP2100/hp2100_sys.c
@@ -1872,7 +1872,7 @@ const HP_WORD odd_parity [256] = {                      /* odd parity table */
 
 char sim_name [] = "HP 2100";                   /* the simulator name */
 
-int32 sim_emax = MAX_INSTR_LENGTH;              /* the maximum number of words in any instruction */
+size_t sim_emax = MAX_INSTR_LENGTH;             /* the maximum number of words in any instruction */
 
 void (*sim_vm_init) (void) = &one_time_init;    /* a pointer to the one-time initializer */
