+++ Intel-Systems/ibmpc/ibmpc_sys.c
@@ -71,7 +71,7 @@ char sim_name[] = "IBM PC";
 
 REG *sim_PC = &i8088_reg[0];
 
-int32 sim_emax = 4;
+size_t sim_emax = 4;
 
 DEVICE *sim_devices[] = {
     &i8088_dev,
