+++ Interdata/id16_sys.c
@@ -66,7 +66,7 @@ char sim_name[] = "Interdata 16b";
 
 REG *sim_PC = &cpu_reg[0];
 
-int32 sim_emax = 2;
+size_t sim_emax = 2;
 
 DEVICE *sim_devices[] = {
     &cpu_dev,
