+++ S3/s3_sys.c
@@ -63,7 +63,7 @@ char sim_name[] = "System/3";
 
 REG *sim_PC = &cpu_reg[0];
 
-int32 sim_emax = 6;
+size_t sim_emax = 6;
 
 DEVICE *sim_devices[] = {
     &cpu_dev,
