+++ scp.h
@@ -404,7 +404,7 @@ extern REG *sim_PC;
 extern const char *sim_stop_messages[SCPE_BASE];
 extern t_stat sim_instr (void);
 extern t_stat sim_load (FILE *ptr, CONST char *cptr, CONST char *fnam, int flag);
-extern int32 sim_emax;
+extern size_t sim_emax;
 extern t_stat fprint_sym (FILE *ofile, t_addr addr, t_value *val,
     UNIT *uptr, int32 sw);
 extern t_stat parse_sym (CONST char *cptr, t_addr addr, UNIT *uptr, t_value *val,
